+++ tester/src/ft_sizettest_for_str.c
@@ -6,7 +6,7 @@
 /*   By: jeportie <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/18 13:56:27 by jeportie          #+#    #+#             */
-/*   Updated: 2023/12/18 14:00:13 by jeportie         ###   ########.fr       */
+/*   Updated: 2023/12/27 13:32:10 by jeportie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void ft_sizettest_for_str(size_t (*libc_func)(), size_t (*ft_func)(), t_tstlst *lst)
 {
-    size_t		result; 
-    size_t		ft_result;
-    char		*value;
+    int		result; 
+    int		ft_result;
+    char	*value;
 
     if (!ft_strncmp("NULL", lst->test_values[0], 4))
         value = NULL;
@@ -35,10 +35,10 @@ void ft_sizettest_for_str(size_t (*libc_func)(), size_t (*ft_func)(), t_tstlst *
         if (ft_result == SEGFAULT)
             printf("SEGFAULT\n\tlibcvalue: ");
         else
-            printf("%zu\n\tlibc_value: ", ft_result);
+            printf("%zu\n\tlibc_value: ", (size_t)ft_result);
         if (result == SEGFAULT)
             printf("SEGFAULT\n");
         else
-            printf("%zu\n", result);
+            printf("%zu\n", (size_t)result);
     }
 }
